requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
selenium==4.15.2
